Added logging, and some checks if upload data to catalog

+++ MiSEQ/miseq_run_metadata.py
@@ -30,11 +30,19 @@ class CollectRunMetadata:
         xml_runInfo = os.path.join(self.run_path, "RunInfo.xml")
         txt_analysisLog = os.path.join(self.run_path, "AnalysisLog.txt")
 
+        if not(os.path.exists(xml_runParameters) and os.path.exists(xml_GenerateFASTQRunStatistics) and os.path.exists(xml_runInfo) and os.path.exists(txt_analysisLog)):
+            return False
+
         run_data = self._find_run_metadata(xml_runParameters, xml_GenerateFASTQRunStatistics, xml_runInfo, txt_analysisLog)
         run_metadata = os.path.join(self.run_path, "run_metadata.json")
 
+        if not (os.path.exists(os.path.join(self.run_path, "run_metadata.json"))):
+            return False
+
         with open(run_metadata, 'w') as outfile:
             json.dump(run_data, outfile, indent=4)
+
+        return True
 
     def _find_run_metadata(self, run_parameters, generate_FASTQ_run_statistics, run_info, analysis_log): #fifth_source
         run_parameters_tree = ET.parse(run_parameters)
